Fix action dtype: use int64 for discrete actions instead of float32
The systematic dataset was creating actions as float32, but discrete action
encoder expects int64. This caused assertion error during training.

+++ eb_jepa/datasets/atari/systematic_dataset.py
@@ -123,8 +123,8 @@ class SystematicAtariDataset(DatasetBase):
         # states: [T, H, W] -> [1, T, H, W] (grayscale channel)
         states = torch.stack(states).unsqueeze(0)
 
-        # actions: [T] -> [1, T] (discrete actions)
-        actions = torch.tensor(actions, dtype=torch.float32).unsqueeze(0)
+        # actions: [T] -> [1, T] (discrete actions as integers)
+        actions = torch.tensor(actions, dtype=torch.int64).unsqueeze(0)
 
         # Create metadata
         metadata = {
